feat(stats): promote winrate charts C+D to real charts, retire mockups
Replaces the experimental mockups tab and the old jitter scatter with
two real character-winrate visualizations on the Stats tab:

- CharacterWinRateLogScatter: X = Rounds played (log), Y = Win Rate %.
  Fixed-size character portraits, 50% reference line. Spreads the
  low-volume cluster cleanly while keeping high-volume detail.

- CharacterWinRateLollipop: characters laid out on a fixed grid,
  X-position determined solely by rounds played (rightmost = most
  played). Per-character round counts rendered along the bottom.
  All portraits same size — no round-weight encoding, position is
  the only volume signal. Stems run from a 50% baseline, colored
  by above/below.

Also updates CasualPage and BracketView to consume the new log
scatter (drop-in replacement for the old CharacterWinRateScatter).

Removes:
- Mockups tab + /mockups route
- All four mockup variants (A/B/C/D) — C and D are now first-class
- The old random-jitter CharacterWinRateScatter
