+++ client/src/features/stats/charts/CharacterPortraitDot.tsx
