+++ client/src/features/stats/charts/index.ts
@@ -1,3 +1,4 @@
-export { default as CharacterWinRateScatter } from "./CharacterWinRateScatter";
+export { default as CharacterWinRateLogScatter } from "./CharacterWinRateLogScatter";
+export { default as CharacterWinRateLollipop } from "./CharacterWinRateLollipop";
 export { default as CharacterWinRateTable } from "./CharacterWinRateTable";
 export { default as PlayerWinRateBar } from "./PlayerWinRateBar";
