--- client/src/features/stats/charts/mockups/index.ts
@@ -1,4 +1,0 @@
-export { default as MockupA_VerticalBubble } from "./MockupA_VerticalBubble";
-export { default as MockupB_Beeswarm } from "./MockupB_Beeswarm";
-export { default as MockupC_TwoAxis } from "./MockupC_TwoAxis";
-export { default as MockupD_Lollipop } from "./MockupD_Lollipop";
